feat(brc20): Add `tw_misc` and `tw_memory::RawPtrTrait` from `dev` branch
