+++ rust/Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "tw_hash",
     "tw_keypair",
     "tw_memory",
+    "tw_misc",
     "tw_move_parser",
     "tw_proto",
     "tw_starknet",
