+++ rust/tw_bitcoin/Cargo.toml
@@ -12,4 +12,5 @@ serde = { version = "1.0.163", features = [ "derive" ] }
 serde_json = "1.0.96"
 tw_encoding = { path = "../tw_encoding" }
 tw_memory = { path = "../tw_memory" }
+tw_misc = { path = "../tw_misc" }
 tw_proto = { path = "../tw_proto" }
